refactor(#38): Server/Client 컴포넌트 분리로 assets 오용 방지
