+++ libs/ui/src/index.ts
@@ -2,5 +2,4 @@ export * from './lib';
 export * from './context';
 export * from './hooks';
 export * from './design';
-
 export * from './api';
