+++ libs/ui/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../tsconfig.base.json",
+  "extends": "../../tsconfig.base.json",
   "files": [],
   "include": [],
   "references": [
@@ -7,4 +7,4 @@
       "path": "./tsconfig.lib.json"
     }
   ]
-}
+}
