+++ tsconfig.base.json
@@ -6,7 +6,8 @@
     "target": "es2022",
     "strict": true,
     "paths": {
-      "@meezy/ui": ["libs/ui/src/index.ts"]
+      "@meezy/ui": ["libs/ui/src/index.ts"],
+      "@meezy/ui/assets": ["libs/ui/src/assets/index.client.ts"]
     }
   }
 }
